Move window creation out of backend
